Externalize utilities
